+++ jscomp/syntax/tests/conversion/reason/recursiveType.ml
@@ -1,4 +1,0 @@
-type tree =
-  < label: string  ;left: tree option  ;right: tree option   >  Js.t
-
-type nonrec t = t