+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Drop Node.js version <18 support, due to it reaching End-of-Life. https://github.com/rescript-lang/rescript-compiler/pull/6429
 - Remove deprecated -bs-super-errors option. https://github.com/rescript-lang/rescript-compiler/pull/6814
 - Some global names and old keywords are no longer prefixed. https://github.com/rescript-lang/rescript-compiler/pull/6831
+- Remove ml parsing tests and conversion from `.ml` to `.res` via format. https://github.com/rescript-lang/rescript-compiler/pull/6848
 
 #### :bug: Bug Fix